+++ src/components/BatchVoting/DefaultOptions/AbstainOptions.tsx
@@ -2,33 +2,14 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 import React from 'react';
-import { Form, FormInstance } from 'antd';
+import { Form } from 'antd';
 import BalanceInput from '~src/ui-components/BalanceInput';
 import BN from 'bn.js';
 import { useTheme } from 'next-themes';
 import styled from 'styled-components';
+import { IAbstainOptions } from '../types';
 
-export enum EFormType {
-	AYE_NAY_FORM = 'aye-nay-form',
-	SPLIT_FORM = 'split-form',
-	ABSTAIN_FORM = 'abstain-form'
-}
-
-interface Props {
-	formName: EFormType;
-	form: FormInstance<any>;
-	handleSubmit: () => void;
-	onBalanceChange: (pre: BN) => void;
-	onAyeValueChange?: (pre: BN) => void;
-	onNayValueChange?: (pre: BN) => void;
-	onAbstainValueChange?: (pre: BN) => void;
-	className?: string;
-	forSpecificPost?: boolean;
-	showConvictionBar?: boolean;
-	isUsedInTinderWebView?: boolean;
-}
-
-const AbstainOptions = ({ form, formName, handleSubmit, onAyeValueChange, onNayValueChange, className }: Props) => {
+const AbstainOptions = ({ form, formName, handleSubmit, onAyeValueChange, onNayValueChange, className }: IAbstainOptions) => {
 	const { resolvedTheme: theme } = useTheme();
 
 	const renderBalanceInput = (label: string, placeholder: string, onChange: (balance: BN) => void, formItemName: string) => (
